Fix some of the failing unit tests.
Adapt some of the existing and failing unit tests to the changed
parameters of Huntercoin.  Not yet updated and still failing are many
name tests.  Also one of the transaction tests is still failing, but
this seems to be related to a wrongly written test itself, see
https://github.com/bitcoin/bitcoin/pull/6863.  Trying to get this merged
upstream instead of a custom fix here.

Furthermore, there are not yet any actual unit tests written newly for
Huntercoin or game logic in general.