Fix name commands for Huntercoin.
Update the RPC name commands name_new, name_firstupdate and name_update
(no name_register yet) to work.  This requires mostly computing and
sending the correct amount of locked coins.

Also update and fix name DB validation as this used to fail on regtest
net when registering names.

Not yet included are regression tests for the name handling in the HUC
context and also the "correct" name_update fee policy.