Fix indexing of game transactions.
Add a regression test for -txindex lookup of game transactions, fix the
way this is done.  We need to keep nFile and nPos to look up the block
header in the block file, as well as store the offset (nPos+nTxOffset)
in the undo file to retrieve the tx itself.  For this, we use a negative
nTxOffset to signify game tx and keep that value.