Properly clean the mempool on name/game conflicts.
Ensure that the mempool is properly cleaned when certain conflicts arise
due to game or name changes.  In particular, transactions need to be
removed from it in either of these situations:

*) A pending name registration when the same name is registered already
   in an incoming block.

*) A pending name update when the hunter being updated is killed.

*) A name registration if the registered name is "revived" due to a
   block disconnect.

All of them are tested with a new regtest and the code is accordingly
fixed to make the test work.