Implement getcoinsnapshot RPC call.
This function can be used to extract the HUC snapshot that will be used
to reward Chimaera.  This method iterates through the UTXO set, and sums
up all coins held in addresses.  Things to note:

  * Coins locked in names are ignored (but the total amount is returned
    for information).
  * Coins in "strange" scripts (all except for pay-to-pubkey and
    pay-to-pubkeyhash) are also ignored.  There seem to be none anyway.
  * Coins in the game world are not considered.

Finally, only addresses with a balance (what counts is the total sum of
all coins in the address, not in individual outputs) of at least a given
minimum amount are returned.

The snapshot is always returned at the current best block, but one can
use 'invalidateblock' to make the client sync to block 1680k.  (And then
'reconsiderblock' to get back to normal.  Probably a backup of the
blockchain is also a good idea.)