Put character states in JSON in own sub-object.
Instead of putting the character states with keys like "0" or "2"
directly into the JSON object of a player state, put them into a
separate "character" object.  This makes it easier to enumerate all
characters without the need to do dirty tricks as before.

Also bump the version to 0.14.99.1 to indicate this change.

This resolves https://github.com/domob1812/huntercore/issues/12.