Fix getstatsforheight test.
Update the getstatsforheight.py regtest for upstream refactorings
and also run it with only one node.  That should be enough and allows
us to get rid of the costly sync'ing.

Also adds the non-"game" regtests that are still relevant for Huntercoin
to the run_game_tests.sh utility script.